Softec Project/fairgig/services/auth/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
email-validator==2.2.0
